aider: Add function to get the number of authors.
# Aider chat conversation:

USER: similar to `getReviewersCount`, implement author count

ASSISTANT: To implement a function similar to `getReviewersCount` for authors, we need to modify the `src/PeerReview.sol` file to add a new function that returns the count of authors. This function will be named `getAuthorsCount`.

Here is the *SEARCH/REPLACE* block for adding the `getAuthorsCount` function:

src/PeerReview.sol
```solidity
<<<<<<< SEARCH
    // Function to get the number of reviewers
    function getReviewersCount() public view returns (uint256) {
        return reviewers.length;
    }
=======
    // Function to get the number of reviewers
    function getReviewersCount() public view returns (uint256) {
        return reviewers.length;
    }

    // Function to get the number of authors
    function getAuthorsCount() public view returns (uint256) {
        return authors.length;
    }
>>>>>>> REPLACE
```

This change adds a new function `getAuthorsCount` that returns the count of authors by returning the length of the `authors` array. This function is similar to the existing `getReviewersCount` function but operates on the `authors` array instead.