aider: Added a test to check that scores are not 0 after running `findReviewers`.
# Aider chat conversation:

USER: add a test to check that scores are not 0 after running `findReviewers`

ASSISTANT: To add a test that checks scores are not 0 after running `findReviewers`, we need to modify the file `test/PeerReview.t.sol`. We will add a new test function that first triggers the `findReviewers` function for a submission and then asserts that the scores of the selected reviewers are not 0.

Here is the *SEARCH/REPLACE block* for adding the new test:

test/PeerReview.t.sol
```solidity
<<<<<<< SEARCH
}
=======
  /// @dev Test that scores are not 0 after running findReviewers
  function testScoresAreNotZeroAfterFindingReviewers() public {
    // Simulate an author submitting a data object
    vm.startPrank(0x70997970C51812dc3A010C7d01b50e0d17dc79C8); // Simulate call from author's address
    string memory question = "why do we need gasless transactions?";
    string memory response = "We need gasless transactions to make blockchain easier to use and access for everyone, especially newcomers, by removing the need for upfront crypto and managing fees. This improves user experience and potentially helps scale the technology. However, it introduces some centralization concerns.";
    uint256 submissionId = peerReview.submitData(question, response);
    vm.stopPrank();

    // Trigger the function to find suitable reviewers for the submission
    peerReview.findReviewers(submissionId);

    // Verify the scores of the selected reviewers are not 0
    address[] memory selectedReviewers = peerReview.getSelectedReviewers(submissionId);
    for (uint256 i = 0; i < selectedReviewers.length; i++) {
      uint256 reviewerScore = peerReview.getReviewerScore(selectedReviewers[i]);
      assertTrue(reviewerScore > 0, "Reviewer score should not be 0");
    }
  }
}
>>>>>>> REPLACE
```

This test function simulates an author submitting a data object, triggers the `findReviewers` function, and then checks that the scores of the selected reviewers are greater than 0. Note that for this test to work, you would need to implement a `getReviewerScore` function in the `PeerReview` contract, which returns the score of a given reviewer. If such a function does not exist, you would need to add it to the contract.